refactor(auth): extract component logic to a custom hook
